Move component under versions folder for 2.0.0
Also added a ParseComponent function which can determine the type of
component to further unmarshal the yaml into
